Moved qp solver under generic. Parallel cpu backend tests don't pass. Investigate this
