+++ tests/lava/lib/optimization/solvers/generic/qp/test_process.py
@@ -6,7 +6,7 @@
 import unittest
 import numpy as np
 
-from lava.lib.optimization.solvers.qp.processes import (
+from lava.lib.optimization.solvers.generic.qp.processes import (
     ConstraintCheck,
     ConstraintNeurons,
     QPDense,
